fix: Vanilla enums can be used custom class variable declarations
WolvenKit would refuse to compile custom class with vanilla enum variables. This is because the compiler was fed an enum class that did not properly extend the vanilla enum class. This is because the vanilla enum calss is in WolvenKit.CR2W and not WolvenKit.CR2W.Types. This is stop gap fix that puts the enum class in the WolvenKit.CR2W namespace however in the future the namespace of the vanilla enums should be corrected.