fix: Try to improve flow diagram performance
The flow diagram in WolvenKit is very slow when viewing big phases as it loops over thousands of chunks everytime the user zoom/pans redrawing them over and over. Windows Forms is not a suitable interface for such as system and it needs to be rebuilt with a package more suitable for this application. In the meantime, I have tried to improve performance by not drawing the connections when the user is panning but results are still not that great.